Separate antialias.py source code file
